Rename DynamicCalibrationConfig to Control in headers
