+++ bindings/python/src/DatatypeBindings.cpp
@@ -36,7 +36,7 @@ void bind_imagealignconfig(pybind11::module& m, void* pCallstack);
 void bind_imageannotations(pybind11::module& m, void* pCallstack);
 #ifdef DEPTHAI_HAVE_DYNAMIC_CALIBRATION_SUPPORT
 void bind_dynamic_calibration_results(pybind11::module& m, void* pCallstack);
-void bind_dynamic_calibration_config(pybind11::module& m, void* pCallstack);
+void bind_dynamic_calibration_control(pybind11::module& m, void* pCallstack);
 #endif  // DEPTHAI_HAVE_DYNAMIC_CALIBRATION_SUPPORT
 
 void DatatypeBindings::addToCallstack(std::deque<StackFunction>& callstack) {
@@ -77,7 +77,7 @@ void DatatypeBindings::addToCallstack(std::deque<StackFunction>& callstack) {
     callstack.push_front(bind_rgbddata);
 #ifdef DEPTHAI_HAVE_DYNAMIC_CALIBRATION_SUPPORT
     callstack.push_front(bind_dynamic_calibration_results);
-    callstack.push_front(bind_dynamic_calibration_config);
+    callstack.push_front(bind_dynamic_calibration_control);
 #endif  // DEPTHAI_HAVE_DYNAMIC_CALIBRATION_SUPPORT
 }
