--- src/pipeline/datatype/DynamicCalibrationConfig.cpp
@@ -1,3 +1,0 @@
-#include "depthai/pipeline/datatype/DynamicCalibrationConfig.hpp"
-
-namespace dai {}  // namespace dai
