+++ src/pipeline/datatype/DynamicCalibrationControl.cpp
@@ -1,0 +1,3 @@
+#include "depthai/pipeline/datatype/DynamicCalibrationControl.hpp"
+
+namespace dai {}  // namespace dai
